refactor: moved all structs to models package
Changes are made to
- RequestBody structs
- Task struct
