+++ backend/handle_tw.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"ccsync_backend/models"
 	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
@@ -11,22 +12,6 @@ import (
 
 	"github.com/google/uuid"
 )
-
-// Task represents a Taskwarrior task
-type Task struct {
-	ID          int32    `json:"id"`
-	Description string   `json:"description"`
-	Project     string   `json:"project"`
-	Tags        []string `json:"tags"`
-	Status      string   `json:"status"`
-	UUID        string   `json:"uuid"`
-	Urgency     float32  `json:"urgency"`
-	Priority    string   `json:"priority"`
-	Due         string   `json:"due"`
-	End         string   `json:"end"`
-	Entry       string   `json:"entry"`
-	Modified    string   `json:"modified"`
-}
 
 // logic to generate client ID for tw config
 func GenerateUUID(email, id string) string {
@@ -42,7 +27,7 @@ func GenerateEncryptionSecret(uuidStr, email, id string) string {
 }
 
 // complete logic (delete config if any->setup config->sync->get tasks->export)
-func FetchTasksFromTaskwarrior(email, encryptionSecret, origin, UUID string) ([]Task, error) {
+func FetchTasksFromTaskwarrior(email, encryptionSecret, origin, UUID string) ([]models.Task, error) {
 	// temporary directory for each user
 	cmd := exec.Command("rm", "-rf", "/root/.task")
 	if err := cmd.Run(); err != nil {
@@ -99,7 +84,7 @@ func SyncTaskwarrior(tempDir string) error {
 }
 
 // export the tasks so as to add them to DB
-func ExportTasks(tempDir string) ([]Task, error) {
+func ExportTasks(tempDir string) ([]models.Task, error) {
 	cmd := exec.Command("task", "export")
 	cmd.Dir = tempDir
 	output, err := cmd.Output()
@@ -108,7 +93,7 @@ func ExportTasks(tempDir string) ([]Task, error) {
 	}
 
 	// Parse the exported tasks
-	var tasks []Task
+	var tasks []models.Task
 	if err := json.Unmarshal(output, &tasks); err != nil {
 		return nil, fmt.Errorf("error parsing tasks: %v", err)
 	}
